Fixing clang compilation warning

+++ source/IOHaptics/test/IOBinaryBands.test.cpp
@@ -296,8 +296,7 @@ TEST_CASE("write/read BandBody on curve") {
     auto byteStreamSize = bitStreamSize % haptics::io::BYTE_SIZE == 0
                               ? bitStreamSize / haptics::io::BYTE_SIZE
                               : (bitStreamSize / haptics::io::BYTE_SIZE) + 1;
-    CHECK(static_cast<int>(std::filesystem::file_size(filename)) ==
-          static_cast<uintmax_t>(byteStreamSize));
+    CHECK(std::filesystem::file_size(filename) == static_cast<uintmax_t>(byteStreamSize));
   }
 
   SECTION("read band body") {
